SIKKA-4387[completed] achieved paste/copy, removed rich-textarea
