+++ docs/documentation/project.json
@@ -1,1 +1,1 @@
-{"generatedAt":1688619728698,"builder":{"name":"webpack5"},"hasCustomBabel":false,"hasCustomWebpack":true,"hasStaticDirs":false,"hasStorybookEslint":false,"refCount":0,"packageManager":{"type":"npm","version":"8.5.0"},"storybookVersion":"6.5.16","language":"typescript","storybookPackages":{"@storybook/addon-actions":{"version":"6.5.16"},"@storybook/addons":{"version":"6.5.16"},"@storybook/builder-webpack5":{"version":"6.5.16"},"@storybook/manager-webpack5":{"version":"6.5.16"},"@storybook/react":{"version":"6.5.16"},"@storybook/storybook-deployer":{"version":"2.8.16"},"@storybook/theming":{"version":"6.5.16"}},"framework":{"name":"react"},"addons":{"storybook-tailwind-dark-mode":{"version":null},"@storybook/addon-links":{"version":"6.5.16"},"@storybook/addon-docs":{"version":"6.5.16"},"@storybook/addon-essentials":{"version":"6.5.16"},"@storybook/addon-postcss":{"options":{"cssLoaderOptions":{"importLoaders":1},"postcssLoaderOptions":{}},"version":"2.0.0"}}}
+{"generatedAt":1688649894757,"builder":{"name":"webpack5"},"hasCustomBabel":false,"hasCustomWebpack":true,"hasStaticDirs":false,"hasStorybookEslint":false,"refCount":0,"packageManager":{"type":"npm","version":"8.5.0"},"storybookVersion":"6.5.16","language":"typescript","storybookPackages":{"@storybook/addon-actions":{"version":"6.5.16"},"@storybook/addons":{"version":"6.5.16"},"@storybook/builder-webpack5":{"version":"6.5.16"},"@storybook/manager-webpack5":{"version":"6.5.16"},"@storybook/react":{"version":"6.5.16"},"@storybook/storybook-deployer":{"version":"2.8.16"},"@storybook/theming":{"version":"6.5.16"}},"framework":{"name":"react"},"addons":{"storybook-tailwind-dark-mode":{"version":null},"@storybook/addon-links":{"version":"6.5.16"},"@storybook/addon-docs":{"version":"6.5.16"},"@storybook/addon-essentials":{"version":"6.5.16"},"@storybook/addon-postcss":{"options":{"cssLoaderOptions":{"importLoaders":1},"postcssLoaderOptions":{}},"version":"2.0.0"}}}
